requirements.txt
absl-py==0.8.1
astor==0.8.1
backports.functools-lru-cache==1.6.1
backports.weakref==1.0.post1
cycler==0.10.0
enum34==1.1.6
funcsigs==1.0.2
functools32==3.2.3.post2
futures==3.3.0
gast==0.2.2
google-pasta==0.1.8
grpcio==1.25.0
h5py==2.10.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
Markdown==3.1.1
matplotlib==2.2.4
mock==3.0.5
numpy==1.16.5
opt-einsum==2.3.2
protobuf==3.11.1
pyparsing==2.4.5
python-dateutil==2.8.1
pytz==2019.3
scipy==1.2.2
six==1.13.0
subprocess32==3.5.4
tensorboard==1.15.0
tensorflow-estimator==1.15.1
tensorflow-gpu==1.15.0
termcolor==1.1.0
Werkzeug==0.16.0
wrapt==1.11.2
